Remove the unused benchmark verdict engine and refocus the roadmap
BenchmarkVerdict shipped in #211 with no caller: 130 lines of library plus
ten tests that nothing in the product invoked. The part of #206 that earns
its place is the coverage warning wired into classify, which catches the
gap that cost four smoke suites on #88. Deleting the rest.

#207 is closed and drops out of the delivery order. Its fix changed 233
lines while fixing nothing reachable -- the rootOnly gap already has a
regression test from #88, and the batch path added in #171 does not touch
the cache at all.

#206 stays, rescoped to the work that is actually left: benchmarks for the
six tools the suite cannot measure, to be added when a change to one of
them needs measuring rather than in advance.

Adds two standing decisions drawn from both mistakes: fix defects that
have a reproduction rather than refactoring latent risk, and do not build
tooling ahead of a caller.

Validation impact: docs.
